+++ src/com/project/my/homeworks/hw5/optional/entities/VipTrip.java
@@ -15,7 +15,7 @@ public class VipTrip {
 		rainyRate = 2;
 	}
 
-	public double calculatePrice(TripParameters tripParameters) {
+	public int calculatePrice(TripParameters tripParameters) {
 		int source = tripParameters.getSource();
 		int destination = tripParameters.getDestination();
 		boolean isRainy = tripParameters.isRainy();
@@ -31,6 +31,6 @@ public class VipTrip {
 			price = x * rainyRate;
 		else
 			price = x;
-		return price;
+		return (int) price;
 	}
 }
